Create README.md
It is a key technology to further reduce the lower limit of detection in Coulter counter by effectively measuring the particle sizes corresponding to the pulse signals annihilated in noise, and these pulse signals are called as low-SNR (signal-to-noise ratio) pulse signals. In this paper, Coulter prototype experiment is set up and innovatively combined with chaotic oscillators to identify particle sizes corresponding to low-SNR pulse signals, and particle size distribution is given. Firstly, the signals satisfying the Gaussian distribution is used to fit the pulse signals measured in the experiment. After being excited by low-SNR fitting signals, it is proposed to identify particle sizes by using the maxima in the synchronization mutation phenomena between oscillators of the chaotic strongly coupled Lorenz-like oscillators, which are used as the independent variables to construct the identification formula of particle sizes. Secondly, the nonlinear circuit of the oscillators is designed and fabricated, which is applied to process pulse signals in the experiment. Finally, selecting the maxima in output signals from the circuit, according to the identification formula contained correction coefficient, particle sizes are obtained and particle size distribution is given. The comparison results with commercial Coulter counter-Multisizer 4e verify that the values of D50 in the distribution converted by the identification formula are effective. The nonlinear circuit can not only remove various types of noise in the experiment, but also effectively process low-SNR pulse signals to obtain reliable particle sizes.